+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/hin/compare/v0.1.4...HEAD)
 ------------------------------------------------------------------------
+### Fixed
+- allow committing child when not the only file
 
 [0.1.4](https://github.com/jshwi/hin/releases/tag/v0.1.4) - 2023-06-17
 ------------------------------------------------------------------------